requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
Click==7.0
cryptography==2.6.1
Flask==1.0.3
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql==0.0.2
mysqlclient==1.4.2
pycparser==2.19
pyOpenSSL==19.0.0
six==1.12.0
SQLAlchemy==1.3.4
Werkzeug==0.15.4